invenio/pyproject.toml
[project]
name = "invenio-ccmm"
version = "0.1.0"
description = "Vocabulary conversion from CCMM to Invenio fixtures"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.2.0",
    "pyyaml>=6.0.2",
    "rdflib[xml]>=7.1.4",
    "tenacity>=9.1.2",
    "tqdm>=4.67.1",
    "types-tqdm>=4.67.0.20250516",
]

[project.scripts]
convert_vocabularies = "invenio_ccmm.convert_vocabularies:convert_vocabularies"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


